Change price field type to double precision
This commit alters the type of the 'price_value' column in the 'products' table of the 'product' schema from decimal (numeric type) to double precision. The type change is carried out using a new EF Core migration.

+++ backend/src/BiteRight.Application/Commands/Products/Create/CreateRequest.cs
@@ -18,7 +18,7 @@ namespace BiteRight.Application.Commands.Products.Create;
 public record CreateRequest(
     string Name,
     string Description,
-    decimal? Price,
+    double? Price,
     Guid? CurrencyId,
     ExpirationDateKindDto ExpirationDateKind,
     DateOnly? ExpirationDate,
